requirements.txt
tensorflow
keras
matplotlib
argparse==1.4.0
numpy==1.24.3
PILLOW